Final commit for tests with detroy tests. This closes #4. Linked commits : e0d818bbb9e1b606100f656ad6b5ef0f4300b72a 9de262cd84461e88c80c7f83cc32784bbc550500 7e59dc6b6c83eafa0e385a95c89a1608496efdfa 0ff76093dc1cdf7d3a73b7f7828dcc689686862a d74a72e8a3674f309d2d355de0d71079e55bb722

+++ src/controllers/playerController.js
@@ -29,7 +29,11 @@ export const playerController = {
         return res.status(404).end();
     },
     destroy(req, res) {
-        playerService.deleteByName(req.params.id);
-        res.status(204).end();
+        var player = playerService.findByName(req.params.id);
+        if (player) {
+            playerService.deleteByName(req.params.id);
+            return res.status(204).end();
+        }
+        return res.status(404).end();
     }
 };
